refactor: snack bar 보여주는 함수 위치 변경
