+++ src/js/utils/dom.js
@@ -3,14 +3,4 @@ const $ = (selector, baseElement = document) =>
 const $$ = (selector, baseElement = document) =>
   baseElement.querySelectorAll(selector);
 
-const showExceptionSnackBar = message => {
-  const $snackBar = $('#snack-bar');
-  $snackBar.classList.add('show');
-  $snackBar.textContent = message;
-
-  setTimeout(() => {
-    $snackBar.classList.remove('show');
-  }, 2000);
-};
-
-export { $, $$, showExceptionSnackBar };
+export { $, $$ };
